refactored comment files
