+++ package/components/inline-comment/floating-comment/index.ts
@@ -1,0 +1,1 @@
+export { CommentFloatingContainer } from './comment-floating-container';
